Diabetes Aid Bot/requirements.txt
# Hugging Face & Gradio
huggingface_hub==0.25.2
gradio==5.7.1

# FAISS for Vector Search
faiss-cpu==1.9.0

# LangChain for Retrieval-Augmented Generation (RAG)
langchain==0.3.15 
langchain-community==0.3.7
langchain-huggingface==0.1.2
langchain-openai==0.2.10

# OpenAI API for LLM Responses
openai==1.59.9


# Data Handling & Processing
numpy==1.26.4
pydantic==2.9.2
python-dotenv==1.0.1

# PDF Processing
PyPDF2  

# Search Functionality (Optional for Expanding Queries)
duckduckgo-search==2.9.0

# Interactive Shell for Debugging (Optional)
ipython==8.29.0

# File Type Detection (Optional)
python-magic==0.4.27






